requirements.txt
python-telegram-bot>=20.7
asyncpg>=0.29.0
pydantic>=2.5.0
google-generativeai>=0.3.0
httpx>=0.25.0
python-dotenv>=1.0.0
nest-asyncio>=1.5.8
psycopg2-binary>=2.9.9
aiohttp>=3.9.0
uvloop>=0.19.0
pydantic-ai>=0.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.10.0
isort>=5.12.0
mypy>=1.7.0 